[libnozzle] add nozzle_get_ips tests and kill nozzle_test
