--- libnozzle/tests/nozzle_test.c
@@ -1,241 +1,0 @@
-/*
- * Copyright (C) 2010-2018 Red Hat, Inc.  All rights reserved.
- *
- * Author: Fabio M. Di Nitto <fabbione@kronosnet.org>
- *
- * This software licensed under GPL-2.0+, LGPL-2.0+
- */
-
-#include "config.h"
-
-#include <string.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <errno.h>
-#include <stdint.h>
-#include <limits.h>
-#include <sys/socket.h>
-
-#include "test-common.h"
-
-char testipv4_1[IPBUFSIZE];
-char testipv4_2[IPBUFSIZE];
-char testipv6_1[IPBUFSIZE];
-char testipv6_2[IPBUFSIZE];
-
-static int check_knet_set_del_ip(void)
-{
-	char device_name[IFNAMSIZ];
-	size_t size = IFNAMSIZ;
-	char verifycmd[2048];
-	int err = 0;
-	nozzle_t nozzle;
-	struct nozzle_ip *ip_list = NULL, *ip_list_tmp = NULL;
-	int ip_list_entries = 0;
-	char *error_string = NULL;
-
-	printf("Testing interface add/remove ip\n");
-
-	memset(device_name, 0, size);
-
-	nozzle = nozzle_open(device_name, size, NULL);
-	if (!nozzle) {
-		printf("Unable to init %s\n", device_name);
-		return -1;
-	}
-
-	printf("Adding ip: %s/24\n", testipv4_1);
-
-	err = nozzle_add_ip(nozzle, testipv4_1, "24");
-	if (err < 0) {
-		printf("Unable to assign IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	printf("Adding ip: %s/24\n", testipv4_2);
-
-	err = nozzle_add_ip(nozzle, testipv4_2, "24");
-	if (err < 0) {
-		printf("Unable to assign IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	printf("Adding duplicate ip: %s/24\n", testipv4_1);
-
-	err = nozzle_add_ip(nozzle, testipv4_1, "24");
-	if (err < 0) {
-		printf("Unable to find IP address in libnozzle db\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	printf("Checking ip: %s/24\n", testipv4_1);
-
-	memset(verifycmd, 0, sizeof(verifycmd));
-	snprintf(verifycmd, sizeof(verifycmd)-1,
-#ifdef KNET_LINUX
-		 "ip addr show dev %s | grep -q %s/24", nozzle->name, testipv4_1);
-#endif
-#ifdef KNET_BSD
-		 "ifconfig %s | grep -q %s", nozzle->name, testipv4_1);
-#endif
-	err = execute_bin_sh_command(verifycmd, &error_string);
-	if (error_string) {
-		printf("Error string: %s\n", error_string);
-		free(error_string);
-		error_string = NULL;
-	}
-	if (err) {
-		printf("Unable to verify IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	printf("Get ip list from libnozzle:\n");
-
-	if (nozzle_get_ips(nozzle, &ip_list) < 0) {
-		printf("Not enough mem?\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	ip_list_tmp = ip_list;
-	ip_list_entries = 0;
-
-	while(ip_list_tmp) {
-		ip_list_entries++;
-		printf("Found IP %s %s in libnozzle db\n", ip_list_tmp->ipaddr, ip_list_tmp->prefix);
-		ip_list_tmp = ip_list_tmp->next;
-	}
-
-	if (ip_list_entries != 2) {
-		printf("Didn't get enough ip back from libnozzle?\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	printf("Deleting ip: %s/24\n", testipv4_1);
-
-	err = nozzle_del_ip(nozzle, testipv4_1, "24");
-	if (err < 0) {
-		printf("Unable to delete IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	printf("Deleting ip: %s/24\n", testipv4_2);
-
-	err = nozzle_del_ip(nozzle, testipv4_2, "24");
-	if (err < 0) {
-		printf("Unable to delete IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	printf("Deleting again ip: %s/24\n", testipv4_1);
-
-	err = nozzle_del_ip(nozzle, testipv4_1, "24");
-	if (err < 0) {
-		printf("Unable to delete IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	memset(verifycmd, 0, sizeof(verifycmd));
-	snprintf(verifycmd, sizeof(verifycmd)-1,
-#ifdef KNET_LINUX
-		 "ip addr show dev %s | grep -q %s/24", nozzle->name, testipv4_1);
-#endif
-#ifdef KNET_BSD
-		 "ifconfig %s | grep -q %s", nozzle->name, testipv4_1);
-#endif
-	err = execute_bin_sh_command(verifycmd, &error_string);
-	if (error_string) {
-		printf("Error string: %s\n", error_string);
-		free(error_string);
-		error_string = NULL;
-	}
-	if (!err) {
-		printf("Unable to verify IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	printf("Adding ip: %s/64\n", testipv6_1);
-
-	err = nozzle_add_ip(nozzle, testipv6_1, "64");
-	if (err < 0) {
-		printf("Unable to assign IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	memset(verifycmd, 0, sizeof(verifycmd));
-	snprintf(verifycmd, sizeof(verifycmd)-1,
-#ifdef KNET_LINUX
-		 "ip addr show dev %s | grep -q %s/64", nozzle->name, testipv6_1);
-#endif
-#ifdef KNET_BSD
-		 "ifconfig %s | grep -q %s", nozzle->name, testipv6_1);
-#endif
-	err = execute_bin_sh_command(verifycmd, &error_string);
-	if (error_string) {
-		printf("Error string: %s\n", error_string);
-		free(error_string);
-		error_string = NULL;
-	}
-	if (err) {
-		printf("Unable to verify IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	printf("Deleting ip: %s/64\n", testipv6_1);
-
-	err = nozzle_del_ip(nozzle, testipv6_1, "64");
-	if (err) {
-		printf("Unable to delete IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-	memset(verifycmd, 0, sizeof(verifycmd));
-	snprintf(verifycmd, sizeof(verifycmd)-1,
-#ifdef KNET_LINUX
-		 "ip addr show dev %s | grep -q %s/64", nozzle->name, testipv6_1);
-#endif
-#ifdef KNET_BSD
-		 "ifconfig %s | grep -q %s", nozzle->name, testipv6_1);
-#endif
-	err = execute_bin_sh_command(verifycmd, &error_string);
-	if (error_string) {
-		printf("Error string: %s\n", error_string);
-		free(error_string);
-		error_string = NULL;
-	}
-	if (!err) {
-		printf("Unable to verify IP address\n");
-		err=-1;
-		goto out_clean;
-	}
-
-out_clean:
-	nozzle_close(nozzle);
-
-	return err;
-}
-
-int main(void)
-{
-	need_root();
-
-	make_local_ips(testipv4_1, testipv4_2, testipv6_1, testipv6_2);
-
-	if (check_knet_set_del_ip() < 0)
-		return -1;
-
-	return 0;
-}
